# web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Application API",
      "description": "API cho các chức năng quản trị và người dùng trong ứng dụng",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8000",
        "description": "Local development server"
      }
    ],
    "paths": {
...
      "/login": {
        "get": {
          "summary": "Hiển thị form đăng nhập",
          "description": "Trả về form đăng nhập cho người dùng",
          "operationId": "showLoginForm",
          "tags": ["Authentication"],
          "responses": {
            "200": {
              "description": "Form đăng nhập",
              "content": {
                "application/json": { "schema": { "type": "object" } }
              }
            }
          }
        },
        "post": {
          "summary": "Xử lý đăng nhập",
          "description": "Xác thực thông tin đăng nhập của người dùng",
          "operationId": "login",
          "tags": ["Authentication"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["email", "password"],
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string", "format": "password" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Đăng nhập thành công",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": { "type": "string", "example": "Đăng nhập thành công!" }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Lỗi xác thực",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "object" }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/reset-password": {
        "post": {
          "summary": "Đặt lại mật khẩu",
          "description": "Xử lý yêu cầu đặt lại mật khẩu",
          "operationId": "resetPassword",
          "tags": ["Authentication"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["email", "password", "password_confirmation", "token"],
                  "properties": {
                    "email": { "type": "string", "format": "email" },
                    "password": { "type": "string", "format": "password" },
                    "password_confirmation": { "type": "string", "format": "password" },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Đặt lại mật khẩu thành công",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": { "type": "string", "example": "Đặt lại mật khẩu thành công!" }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Lỗi xác thực",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "object" }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/predict": {
        "post": {
          "summary": "Dự đoán hình ảnh gốm",
          "description": "Xử lý hình ảnh để dự đoán loại gốm",
          "operationId": "predict",
          "tags": ["Prediction"],
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "required": ["image"],
                  "properties": {
                    "image": { "type": "string", "format": "binary", "description": "Tệp hình ảnh gốm" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Kết quả dự đoán",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "prediction": { "type": "object" }
                    }
                  }
                }
              }
            },
            "422": {
              "description": "Lỗi xác thực",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "object" }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }